Refactor installation procedure and docker process using pipenv
