+++ CHANGELOG.md
@@ -9,7 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replaced bash script with python
 - Moved sources in `src` installation folder
 - Corrected IGInterface numpy dependency
-- Enabled --install-dep from install path
 - Added Pipenv integration
 
 ## [1.0.1] - 2019-05-09
